+++ src/robotsim/model/ProductionArea.java
@@ -6,7 +6,7 @@ public class ProductionArea extends Component {
 	
 	private ArrayList<Washer> materials;
 
-	public ProductionArea(double x, double y, int width, int height, ArrayList<Washer> materials) {
+	public ProductionArea(int x, int y, int width, int height, ArrayList<Washer> materials) {
 		super(x, y, new Rectangle(width, height));
 		this.materials = materials;
 	}
